SONARPHP-1330 Serialize and cache the project level symbol table, improve test coverage
